get call broken, will never return node data
get calls (not just this one!) need buffer/memory controll. ZK will not malloc for us. Without this change znode data cannot be read out.

This change only allows up to 2K data to be read, but a znode can have up to 1MB of data, so it probably needs a more clever mechanism.